requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
streamlit
duckduckgo-search
arxiv
wikipedia
bs4
validators
pypdf
pymupdf
unstructured
pytube
sentence_transformers
huggingface_hub
pandas
numpy
SQLAlchemy
mysql-connector-python
numexpr
python-dotenv
ipykernel
pydantic
